requirements.txt
cffi==1.16.0
cryptography==42.0.5
dnspython==2.6.1
netaddr==1.2.1
pycparser==2.21
pymongo==4.6.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
six==1.16.0
